adding create-reservation functionality, refactoring
