+++ components/admin_real/desktop/ReservationTable.jsx
@@ -7,6 +7,7 @@ import {
   calculateRange,
   fetchReservations,
 } from "../shared/reservationTableUtils";
+import CreateReservationButton from "./CreateReservationButton";
 
 export default function ReservationTable() {
   const [data, setData] = React.useState([]);
@@ -31,6 +32,7 @@ export default function ReservationTable() {
   return (
     <div className="bg-white">
       {/* https://mui.com/x/api/data-grid/data-grid/ */}
+      <CreateReservationButton />
       <DataGrid
         rows={data}
         columns={RESERVATION_TABLE_COLUMNS_DESKTOP}
