+++ components/admin_real/shared/reservationTableUtils.js
@@ -1,6 +1,4 @@
 import supabase from "../../../utils/supabase";
-import { Edit } from "@mui/icons-material";
-import { Button } from "@mui/material";
 import DeleteReservationButton from "../desktop/DeleteReservationButton";
 import EditReservationButton from "../desktop/EditReservationButton";
 
@@ -49,6 +47,22 @@ export const updateReservation = async (props) => {
       },
     })
     .eq("id", props.id);
+};
+//Todo: need to add the dive shop id here?
+export const createReservation = async (props) => {
+  return await supabase.from("reservations").insert({
+    date: props.date,
+    time: props.time,
+    diver_certified: props.certified,
+    reservation_type: props.reservationType,
+    number_of_dives: props.experience,
+    diver_information: {
+      name: props.diverInformation.name,
+      lastName: props.diverInformation.lastName,
+      age: props.diverInformation.age,
+      email: props.diverInformation.email,
+    },
+  });
 };
 
 export const DEFAULT_ROW_PROPS = {
